[sftp] Rename file to follow type
